\name{quotenames}
\alias{quotenames}
\title{Quote a "list" of names :}
\usage{
  quotenames(...)
}
\description{
  quotenames(ident,treatment,diagnose) will give a
  character vector = c("ident","treatment","diagnose")
}
\details{
  very useful to quote a large list of names that are not
  quoted (param names, var names, etc)
}

