\name{txt.perc}
\alias{txt.perc}
\title{Numeric to percent with percent symbols for labelling}
\usage{
  txt.perc(x, dec)
}
\description{
  numeric to percents with percent symbols for labelling ,
  with "dec" = decimals to display.
}
\examples{
pp <- c(23.121 , 1.34, 0.345);
fpp <- txt.perc(pp,2); fpp
fpp <- txt.perc(pp,1); fpps
}

